Citation Panel separated
