+++ code/frontend/src/pages/chat/Chat.module.css
@@ -157,90 +157,6 @@
     flex-grow: 0;
 }
 
-.citationPanel {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 16px 16px;
-    gap: 8px;
-    background: #FFFFFF;
-    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.14), 0px 0px 2px rgba(0, 0, 0, 0.12);
-    border-radius: 8px;
-    flex: auto;
-    order: 0;
-    align-self: stretch;
-    flex-grow: 0.3;
-    max-width: 30%;
-    overflow-y: scroll;
-    max-height: calc(100vh - 100px);
-}
-
-.citationPanelHeaderContainer {
-    width: 100%;
-}
-
-.citationPanelHeader {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 600;
-    font-size: 18px;
-    line-height: 24px;
-    color: #000000;
-    flex: none;
-    order: 0;
-    flex-grow: 0;
-}
-
-.citationPanelDismiss {
-    width: 18px;
-    height: 18px;
-    color: #424242;
-}
-
-.citationPanelDismiss:hover {
-    background-color: #D1D1D1;
-    cursor: pointer;
-}
-
-.citationPanelTitle {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 22px;
-    color: #323130;
-    margin-top: 12px;
-    margin-bottom: 12px;
-}
-
-.citationPanelContent {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 20px;
-    color: #000000;
-    flex: none;
-    order: 1;
-    align-self: stretch;
-    flex-grow: 0;
-}
-
-.citationPanelDisclaimer {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 400;
-    font-size: 12px;
-    display: flex;
-    color: #707070;
-}
-
-.citationPanelContent h1 {
-
-    line-height: 30px;
-
-}
-
 .historyPanelTopRightButtons {
     height: 48px;
 }
@@ -309,13 +225,6 @@
   /* High contrast mode specific styles */
   @media screen and (-ms-high-contrast: active), (forced-colors: active) {
     .chatContainer, .chatMessageUserMessage {
-        border: 2px solid WindowText;
-        padding: 10px;
-        background-color: Window;
-        color: WindowText;
-    }
-
-    .citationPanel , .citationPanelHeader, .citationPanelTitle, .citationPanelContent{
         border: 2px solid WindowText;
         padding: 10px;
         background-color: Window;
